requirements.txt
spotipy==2.4.4
youtube-dl==2017.7.2
google-api-python-client==1.6.2
SpeechRecognition==3.7.1
PyAudio==0.2.11
pygame==1.9.3

python-vlc==1.1.2